pyproject.toml
[project]
name = "text2sql"
version = "0.0.1"
description = "Text To SQL"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pathlib>=1.0.1",
    "psycopg2-binary>=2.9.10",
    "pydantic-ai>=0.0.20",
    "pydantic-ai-slim[openai]>=0.0.20",
    "sqlalchemy>=2.0.36",
]

[[tool.uv.index]]
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"
default = true
